+++ GameServerLib/Items/InventoryManager.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using GameServerCore.Domain;
-using LeagueSandbox.GameServer.Content;
 
 namespace LeagueSandbox.GameServer.Items
 {
@@ -25,12 +24,12 @@ namespace LeagueSandbox.GameServer.Items
             return _inventory.SetExtraItem(slot, item);
         }
 
-        public Item GetItem(int slot)
+        public Item GetItem(byte slot)
         {
             return _inventory.GetItem(slot);
         }
 
-        public void RemoveItem(int slot)
+        public void RemoveItem(byte slot)
         {
             _inventory.RemoveItem(slot);
         }
@@ -42,10 +41,10 @@ namespace LeagueSandbox.GameServer.Items
 
         public byte GetItemSlot(Item item)
         {
-            return (byte)_inventory.GetItemSlot(item);
+            return _inventory.GetItemSlot(item);
         }
 
-        public void SwapItems(int slot1, int slot2)
+        public void SwapItems(byte slot1, byte slot2)
         {
             _inventory.SwapItems(slot1, slot2);
         }
@@ -101,7 +100,7 @@ namespace LeagueSandbox.GameServer.Items
             return SetExtraItem(slot, (ItemType)item);
         }
 
-        IItem IInventoryManager.GetItem(int slot)
+        IItem IInventoryManager.GetItem(byte slot)
         {
             return GetItem(slot);
         }
